:recycle: Move can_connect helper to test utilities
